requirements.txt
python-nmap==0.7.1
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
APScheduler==3.10.4
python-dotenv==1.0.1
waitress==3.0.0 